requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.4.0
beautifulsoup4==4.11.1
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.12
colorama==0.4.4
cycler==0.11.0
Deprecated==1.2.13
fonttools==4.33.3
idna==3.3
kiwisolver==1.4.2
lxml==4.8.0
matplotlib==3.5.1
MLB-StatsAPI==1.4.2
multidict==6.0.2
numpy==1.22.3
packaging==21.3
pandas==1.4.2
Pillow==9.1.0
py-cord==2.0.0b7
pyarrow==7.0.0
pybaseball==2.2.1
pycparser==2.21
PyGithub==1.55
PyJWT==2.3.0
PyNaCl==1.5.0
pyparsing==3.0.8
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
requests==2.27.1
scipy==1.8.0
six==1.16.0
soupsieve==2.3.2.post1
tqdm==4.64.0
typing_extensions==4.2.0
urllib3==1.26.9
wrapt==1.14.0
yarl==1.7.2